+++ src/ast.rs
@@ -454,6 +454,7 @@ pub enum Pattern<Constructor> {
         args: Vec<CallArg<Self>>,
         module: Option<String>,
         constructor: Constructor,
+        with_spread: bool,
     },
 
     Tuple {